Extracts a client component from Trick Detail Page
